+++ SEPPmail365cloud.psm1
@@ -7,10 +7,16 @@ $InteractiveSession = [System.Environment]::UserInteractive
 Write-Verbose 'Request terminating errors by default'
 $PSDefaultParameterValues['*:ErrorAction'] = [System.Management.Automation.ActionPreference]::Stop
 
+Write-Verbose 'Initialize argument completer scriptblocks'
+$paramDomSB = {
+    # Read Accepted Domains for domain selection
+    Get-AcceptedDomain -Erroraction silentlycontinue|select-Object -ExpandProperty DomainName
+}
+
+
 Write-Verbose 'Loading Module Files'
 . $ModulePath\Private\PrivateFunctions.ps1
 . $ModulePath\Private\ConfigBundle.ps1
-#. $ModulePath\Private\SetupTypes.ps1
 . $ModulePath\Public\Common.ps1
 . $ModulePath\Public\Rules.ps1
 . $ModulePath\Public\Connectors.ps1